+++ apps/member-profile/app/routes/_profile.home.activation.tsx
@@ -242,28 +242,6 @@ function ActivationList() {
       </ActivationItem>
 
       <ActivationItem
-        label="Open a Weekly Newsletter"
-        requirement="open_email_campaign"
-      >
-        <ActivationItem.Description>
-          You are automatically subscribed to our weekly newsletter -- new
-          issues are typically sent on Wednesdays. If you haven't received a
-          newsletter after 2 weeks of being in ColorStack, let us know.
-        </ActivationItem.Description>
-
-        <ActivationItem.QuestionList>
-          <ActivationItem.Question
-            question="Where can I find the newsletters?"
-            answer='You will receive the newsletter via email. The subject typically starts with "[ColorStack Newsletter]".'
-          />
-          <ActivationItem.Question
-            question="I opened the newsletter, but it has not been marked as opened. What should I do?"
-            answer='Please allow up to 48 hours for our system to register you opening a newsletter. Sometimes Mailchimp does not register opens properly, so you can guarantee it registers by clicking the button at the bottom of the email that says "Click Here for Activation Checkmark".'
-          />
-        </ActivationItem.QuestionList>
-      </ActivationItem>
-
-      <ActivationItem
         label={
           <>
             Introduce Yourself in{' '}
