+++ packages/core/src/api.ts
@@ -13,7 +13,7 @@ import { oneTimeCodeWorker } from './modules/authentication/one-time-code.worker
 import { eventWorker } from './modules/event/event.worker';
 import { feedWorker } from './modules/feed/feed';
 import { gamificationWorker } from './modules/gamification/gamification.core';
-import { emailMarketingWorker } from './modules/mailchimp/email-marketing.worker';
+import { mailchimpWorker } from './modules/mailchimp/mailchimp';
 import { memberEmailWorker } from './modules/member/member-email.worker';
 import { memberWorker } from './modules/member/member.worker';
 import { profileWorker } from './modules/member/profile.worker';
@@ -37,10 +37,10 @@ export { Environment } from './shared/types';
 export function startBullWorkers(): void {
   airtableWorker.run();
   applicationWorker.run();
-  emailMarketingWorker.run();
   eventWorker.run();
   feedWorker.run();
   gamificationWorker.run();
+  mailchimpWorker.run();
   memberWorker.run();
   memberEmailWorker.run();
   notificationWorker.run();
